Separate bundle module into one file per class
