+++ core/bundle/wrappers/BundleWrapper.py
@@ -1,0 +1,9 @@
+from core.bundle import _Bundle
+
+
+class BundleWrapper(_Bundle):
+    def __init__(self, bundle):
+        self.__class__ = type(
+            bundle.__class__.__name__, (self.__class__, bundle.__class__), {}
+        )
+        self.__dict__ = bundle.__dict__
