+++ core/bundle/wrappers/__init__.py
@@ -1,0 +1,2 @@
+from .BundleWrapper import BundleWrapper
+from .PipedTaskBundleWrapper import PipedTaskBundleWrapper
